recommendation/requirements.txt
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow

fastapi
uvicorn
weaviate-client>=3.26.7,<4.0.0
python-multipart
